+++ CHANGELOG.md
@@ -25,7 +25,9 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Frontend: update `browser-sync` to version `2.17.5` from `2.11.2`.
 - Frontend: update `mkdirp` to version `0.5.1` from `0.3.0`.
 - Fixed broken `manage.py check` command when using `cfgov.settings.test`.
+- Fixed conference registration form capacity logic.
 - Update `snyk` to version `1.19.1` from `1.13.2`.
+- Disable logging below CRITICAL when running Python unit tests.
 - Fixed empty `heading` value in link blobs
 
 ### Removed